refactor: rename `annotations` package to `annotation`
