+++ dwcj-addons/dwcj-googlecharts/src/main/java/org/dwcj/addons/googlecharts/events/GoogleChartReadyEvent.java
@@ -3,8 +3,8 @@ package org.dwcj.addons.googlecharts.events;
 import java.util.Map;
 import org.dwcj.addons.googlecharts.GoogleChart;
 import org.dwcj.component.event.Event;
-import org.dwcj.component.webcomponent.annotations.EventExpressions;
-import org.dwcj.component.webcomponent.annotations.EventName;
+import org.dwcj.component.webcomponent.annotation.EventExpressions;
+import org.dwcj.component.webcomponent.annotation.EventName;
 
 /**
  * Most charts are rendered asynchronously; all Google charts throw a ready event after they have
