+++ dwcj-engine/src/main/java/org/dwcj/component/applayout/AppLayout.java
@@ -9,8 +9,8 @@ import org.dwcj.component.applayout.event.AppLayoutDrawerOpenEvent;
 import org.dwcj.component.event.EventListener;
 import org.dwcj.component.webcomponent.PropertyDescriptor;
 import org.dwcj.component.webcomponent.WebComponent;
-import org.dwcj.component.webcomponent.annotations.NodeAttribute;
-import org.dwcj.component.webcomponent.annotations.NodeName;
+import org.dwcj.component.webcomponent.annotation.NodeAttribute;
+import org.dwcj.component.webcomponent.annotation.NodeName;
 import org.dwcj.component.window.Panel;
 
 /**
