+++ dwcj-engine/src/main/java/org/dwcj/component/applayout/event/AppLayoutDrawerOpenEvent.java
@@ -3,8 +3,8 @@ package org.dwcj.component.applayout.event;
 import java.util.Map;
 import org.dwcj.component.applayout.AppLayout;
 import org.dwcj.component.event.Event;
-import org.dwcj.component.webcomponent.annotations.EventExpressions;
-import org.dwcj.component.webcomponent.annotations.EventName;
+import org.dwcj.component.webcomponent.annotation.EventExpressions;
+import org.dwcj.component.webcomponent.annotation.EventName;
 
 /**
  * Emitted when the drawer is closed.
