+++ dwcj-engine/src/main/java/org/dwcj/component/drawer/event/DrawerCloseEvent.java
@@ -3,8 +3,8 @@ package org.dwcj.component.drawer.event;
 import java.util.Map;
 import org.dwcj.component.drawer.Drawer;
 import org.dwcj.component.event.Event;
-import org.dwcj.component.webcomponent.annotations.EventExpressions;
-import org.dwcj.component.webcomponent.annotations.EventName;
+import org.dwcj.component.webcomponent.annotation.EventExpressions;
+import org.dwcj.component.webcomponent.annotation.EventName;
 
 /**
  * Emitted when the drawer is opened.
