+++ dwcj-engine/src/main/java/org/dwcj/component/webcomponent/WebComponent.java
@@ -29,14 +29,14 @@ import org.dwcj.component.event.EventDispatcher;
 import org.dwcj.component.event.EventListener;
 import org.dwcj.component.htmlcontainer.HtmlContainer;
 import org.dwcj.component.htmlcontainer.event.HtmlContainerJavascriptEvent;
-import org.dwcj.component.webcomponent.annotations.EventExpressions;
-import org.dwcj.component.webcomponent.annotations.EventName;
-import org.dwcj.component.webcomponent.annotations.HtmlViewAttribute;
-import org.dwcj.component.webcomponent.annotations.HtmlViewClassName;
-import org.dwcj.component.webcomponent.annotations.NodeAttribute;
-import org.dwcj.component.webcomponent.annotations.NodeClassName;
-import org.dwcj.component.webcomponent.annotations.NodeName;
-import org.dwcj.component.webcomponent.annotations.NodeProperty;
+import org.dwcj.component.webcomponent.annotation.EventExpressions;
+import org.dwcj.component.webcomponent.annotation.EventName;
+import org.dwcj.component.webcomponent.annotation.HtmlViewAttribute;
+import org.dwcj.component.webcomponent.annotation.HtmlViewClassName;
+import org.dwcj.component.webcomponent.annotation.NodeAttribute;
+import org.dwcj.component.webcomponent.annotation.NodeClassName;
+import org.dwcj.component.webcomponent.annotation.NodeName;
+import org.dwcj.component.webcomponent.annotation.NodeProperty;
 import org.dwcj.component.window.AbstractWindow;
 import org.dwcj.environment.ObjectTable;
 import org.dwcj.exceptions.DwcjComponentDestroyed;
