+++ dwcj-engine/src/main/java/org/dwcj/component/webcomponent/annotation/NodeAttribute.java
@@ -1,4 +1,4 @@
-package org.dwcj.component.webcomponent.annotations;
+package org.dwcj.component.webcomponent.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -9,17 +9,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * The HtmlViewAttribute annotation is used to set attributes on the html view hosting the web
- * component upon creation.
+ * The NodeAttribute annotation is used to set attributes on the web component upon creation.
  *
  * <p>
- * For example, the following annotation will set the attribute "disabled" with value "true" on the
- * html view hosting the web component when it is created.
+ * For example, the following annotation will set the "disabled" attribute to "true" on the web
+ * component when it is created.
  * </p>
  *
  * <pre>
  * {@code
- * &#64;HtmlViewAttribute(name = "disabled", value = "true")
+ * &#64;NodeAttribute(name = "disabled", value = "true")
  * }
  * </pre>
  *
@@ -27,10 +26,10 @@ import java.lang.annotation.Target;
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@Repeatable(HtmlViewAttribute.Container.class)
+@Repeatable(NodeAttribute.Container.class)
 @Inherited
 @Documented
-public @interface HtmlViewAttribute {
+public @interface NodeAttribute {
   /**
    * The name of the attribute.
    *
@@ -46,9 +45,9 @@ public @interface HtmlViewAttribute {
   String value() default "";
 
   /**
-   * The container annotation for {@link HtmlViewAttribute} annotation.
+   * The container annotation for {@link NodeAttribute} annotation.
    *
-   * @see HtmlViewAttribute
+   * @see NodeAttribute
    * @author Hyyan Abo Fakher
    */
   @Target(ElementType.TYPE)
@@ -57,10 +56,10 @@ public @interface HtmlViewAttribute {
   @Documented
   public @interface Container {
     /**
-     * An array of {@link HtmlViewAttribute} annotations.
+     * An array of {@link NodeAttribute} annotations.
      *
-     * @return an array of {@link HtmlViewAttribute} annotations
+     * @return an array of {@link NodeAttribute} annotations
      */
-    HtmlViewAttribute[] value();
+    NodeAttribute[] value();
   }
 }
